Merge multiple .csv files based on a single identifier
Convert all of the excel sheets to .csv. List the file names in the datasheets vector. Use the join_all function from plyr to merge them all (outer merge). Right now the merge variable is 'id_visit', and the data is additionally subset after merging. These are MRS specific, as this is an example, but it can be changed to any dataset.